+++ lib/services/safety/base_stateless.dart
@@ -1,16 +1,23 @@
 import 'package:flutter/material.dart';
 import 'package:nft/services/app/dynamic_size.dart';
+import 'package:nft/utils/app_theme.dart';
 
 /// Remember call super.build(context) in widget
+// ignore: must_be_immutable
 abstract class BaseStateless extends StatelessWidget with DynamicSize {
-  const BaseStateless({Key key}) : super(key: key);
+  BaseStateless({Key key}) : super(key: key);
+
+  AppTheme appTheme;
 
   // Context valid to create providers
+  @mustCallSuper
   @protected
-  void initDependencies(BuildContext context);
+  void initDependencies(BuildContext context) {
+    appTheme = context.theme();
+  }
 
   @protected
-  void afterFirstBuild(BuildContext context);
+  void afterFirstBuild(BuildContext context) {}
 
   @mustCallSuper
   @override