+++ lib/utils/app_route.dart
@@ -9,27 +9,27 @@ class AppRoute {
 
   AppRoute._private();
 
-  //#region ROUTE NAMES
-  // -----------------
+  ///#region ROUTE NAMES
+  /// -----------------
   static const String routeRoot = '/';
   static const String routeHome = '/home';
   static const String routeLogin = '/login';
   static const String routeCounter = '/counter';
 
-  //#endregion
+  ///#endregion
 
   static final AppRoute _instance = AppRoute._private();
 
   static AppRoute get I => _instance;
 
-  // App route observer
+  /// App route observer
   final RouteObserver<Route<dynamic>> routeObserver =
       RouteObserver<Route<dynamic>>();
 
   /// App global navigator key
   final GlobalKey<NavigatorState> navigatorKey = GlobalKey<NavigatorState>();
 
-  // Get app context
+  /// Get app context
   BuildContext get appContext => navigatorKey.currentContext;
 
   /// Generate route for app here
@@ -55,7 +55,7 @@ class AppRoute {
 
 extension AppRouteExt on BuildContext {
   AppRoute route() {
-    return Provider.of<AppRoute>(this, listen: false);
+    return Provider.of(this, listen: false);
   }
 
   NavigatorState navigator() {